Verify that code works as expected
We now verify that everything behaves as normal when the user writes in the three input fields (quote, author, source) and submits it, hence the .include method.
For e.g., taking the first line under the 'verify' comment: does the string returned from the tag with id=quotes return the string attached to the quote variable? If yes, then 
'npm test' will be 'clean'.